+++ packages/plugin-compiler/src/test/kotlin/io/realm/compiler/GenerationExtensionTest.kt
@@ -176,12 +176,12 @@ class GenerationExtensionTest {
         assertTrue(companionObject is RealmObjectCompanion)
 
         val table = companionObject.`$realm$schema`()
-        val excludedPropertiesCount = companionObject.`$realm$excludedProperties`!!.count()
+        val realmFields = companionObject.`$realm$fields`!!
 
         assertEquals("Sample", table.name)
         assertEquals("id", table.primaryKey)
         assertEquals(setOf(ClassFlag.RLM_CLASS_NORMAL), table.flags)
-        assertEquals(sampleModel::class.declaredMemberProperties.size - excludedPropertiesCount, table.properties.size)
+        assertEquals(realmFields.count(), table.properties.size)
         val properties = mapOf(
             // Primary key
             "id" to PropertyType.RLM_PROPERTY_TYPE_INT,